Restructures Routes & Curriculum
