+++ client/src/components/curriculum/CertAccordion.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
 
 .accordion 
     border: 1px solid $border-color
