+++ client/src/components/curriculum/CertChallenge.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
 
 .cert-challenge
     padding: 10px 10px 48px 32px
