+++ client/src/components/curriculum/CertSection.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
     
 .accordian-section 
     position: relative
