+++ client/src/components/curriculum/Certificate.sass
@@ -1,4 +1,4 @@
-@import '../vars.sass'
+@import '../../vars.sass'
 
 .accordion-cert
     position: relative
